api_gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0             # Cliente HTTP async para chamar outros serviços
python-dotenv==1.0.1
pydantic[email]==2.7.1